+++ PMM.Domain/Entities/ProjectAssignment.cs
@@ -14,7 +14,7 @@ namespace PMM.Domain.Entities
 
         public DateOnly? StartedAt { get; set; }
         public DateOnly? EndAt { get; set; }
-        public int? ExpectedHours { get; set; }
-        public int? SpentHours { get; set; }
+        public decimal? ExpectedHours { get; set; }
+        public decimal? SpentHours { get; set; }
     }
 }
